+++ src/interfaces/sqlite/mod.rs
@@ -101,18 +101,9 @@ pub async fn search(
         .await?;
     let mut db_uids = Vec::with_capacity(results.len());
     for (_, indexed_values) in results {
-        for iv in indexed_values {
-            let db_uid = match iv {
-                IndexedValue::Location(db_uid_location) => {
-                    String::from_utf8(db_uid_location.into())
-                        .map_err(|e| FindexErr::ConversionError(format!("Invalid location: {e}")))?
-                }
-                IndexedValue::NextKeyword(_) => {
-                    return Err(FindexErr::Other(
-                        "There should be not newt words".to_string(),
-                    ));
-                }
-            };
+        for location in indexed_values {
+            let db_uid = String::from_utf8(location.into())
+                .map_err(|e| FindexErr::ConversionError(format!("Invalid location: {e}")))?;
             db_uids.push(db_uid);
         }
     }