+++ CHANGELOG.md
@@ -6,12 +6,11 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
-- Change `progress_callback` to return NextKeyword and `search` to return only Location
-- Update `search` to return Location bytes instead of IndexedValues
+- Change `progress_callback` to return `NextKeyword` and `search` to return only `Location`
 
 ### Testing
 
-- Add progress_callback tests
+- Add `progress_callback` tests
 
 ## [2.0.0] - 2023-01-13